Move NI-patches to their own file
