+++ MWSE/PatchNI.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace mwse::patch::ni {
+	void install();
+	void installPostLua();
+}
